feat: Sum of digits
Description:
This pull request adds a new java file which calculates the sum of digits of a number given input by the user. Details:
This pull request introduces a new java program, SumDigits.java that prints the sum of the digits of any number. The purpose of this adding this file is to demonstrate looping concept in java and how digits can be extracted from a number using loop concept. Changes Made:
Added a java file named SumDigits.java in the project directory and calculated the sum of digits implementing loop concept in java programming language.

Checklist:

Output tested in Java IDE, BlueJ. No issues detected. It produces the intended output.
My code follows the guidelines of this project.
My follows the rules of Java programming.
Screenshots: