feat : updated SumOfDigits
# Description
This pull request introduces a new java program, SumDigits.java that prints the sum of the digits of any number. The purpose of this adding this file is to demonstrate looping concept in java and how digits can be extracted from a number using loop concept. 

## Type of change

- [ ] Added a java file named SumDigits.java in the project directory and calculated the sum of digits implementing loop concept in java programming language.
- [ ] This change requires a documentation update

# How Has This Been Tested?

Output tested in Java IDE, BlueJ. 
No issues detected. 
It produces the intended output.
 My code follows the guidelines of this project.
My follows the rules of Java programming.

# Checklist:

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings